+++ comb_spec_searcher/specification.py
@@ -4,11 +4,13 @@ where each of the bi appear exactly once on the left hand side of some rule.
 """
 import logging
 from copy import copy
-from typing import Dict, Generic, Iterable, Iterator, Sequence, Tuple
+from functools import reduce
+from operator import mul
+from typing import Dict, Generic, Iterable, Iterator, List, Sequence, Tuple
 
 import sympy
 from logzero import logger
-from sympy import Eq, Expr, Function, solve, var
+from sympy import Eq, Expr, Function, Number, solve, var
 
 from .combinatorial_class import (
     CombinatorialClass,
@@ -35,7 +37,13 @@ from .strategies import (
     VerificationStrategy,
 )
 from .strategies.rule import AbstractRule
-from .utils import DisableLogging, RecursionLimit, maple_equations, taylor_expand
+from .utils import (
+    DisableLogging,
+    RecursionLimit,
+    maple_equations,
+    pretty_print_equations,
+    taylor_expand,
+)
 
 __all__ = ("CombinatorialSpecification",)
 
@@ -47,6 +55,9 @@ class CombinatorialSpecification(
     A combinatorial specification is a set rules of the form a -> b1, ..., bk
     where each of the bi appear exactly once on the left hand side of some
     rule.
+
+    The default is to expand verified classes, but this can be turned off by
+    setting expand_verified to False.
     """
 
     def __init__(
@@ -59,10 +70,11 @@ class CombinatorialSpecification(
             ]
         ],
         equivalence_paths: Iterable[Sequence[CombinatorialClassType]],
+        expand_verified: bool = True,
     ):
         self.root = root
         self.rules_dict: Dict[CombinatorialClassType, AbstractRule] = {}
-        self._populate_rules_dict(strategies, equivalence_paths)
+        self._populate_rules_dict(strategies, equivalence_paths, expand_verified)
         for rule in list(
             self.rules_dict.values()
         ):  # list as we lazily assign empty rules
@@ -78,6 +90,7 @@ class CombinatorialSpecification(
             ]
         ],
         equivalence_paths: Iterable[Sequence[CombinatorialClassType]],
+        expand_verified: bool,
     ) -> None:
         equivalence_rules: Dict[
             Tuple[CombinatorialClassType, CombinatorialClassType], Rule
@@ -88,20 +101,20 @@ class CombinatorialSpecification(
                 continue
             rule = strategy(comb_class)
             non_empty_children = rule.non_empty_children()
-            if (
-                isinstance(rule, Rule)
-                and len(non_empty_children) == 1
-                and rule.constructor.is_equivalence()
-            ):
+            if rule.is_equivalence():
+                assert isinstance(rule, Rule)
                 equivalence_rules[(comb_class, non_empty_children[0])] = (
                     rule if len(rule.children) == 1 else rule.to_equivalence_rule()
                 )
             elif non_empty_children:
                 self.rules_dict[comb_class] = rule
             elif isinstance(rule, VerificationRule):
-                try:
-                    verification_packs[comb_class] = rule.pack()
-                except InvalidOperationError:
+                if expand_verified:
+                    try:
+                        verification_packs[comb_class] = rule.pack()
+                    except InvalidOperationError:
+                        self.rules_dict[comb_class] = strategy(comb_class)
+                else:
                     self.rules_dict[comb_class] = strategy(comb_class)
             else:
                 raise ValueError("Non verification rule has no children.")
@@ -140,6 +153,7 @@ class CombinatorialSpecification(
                     AlreadyVerified(self.rules_dict), apply_first=True
                 ),
             )
+            logger.info(css.run_information())
             while True:
                 try:
                     css.do_level()
@@ -164,7 +178,7 @@ class CombinatorialSpecification(
             comb_class_rules = [
                 (css.classdb.get_class(label), rule) for label, rule in rules
             ]
-            self._populate_rules_dict(comb_class_rules, comb_class_eqv_paths)
+            self._populate_rules_dict(comb_class_rules, comb_class_eqv_paths, True)
 
     def get_rule(
         self, comb_class: CombinatorialClassType
@@ -224,16 +238,41 @@ class CombinatorialSpecification(
                     yield eq
             except NotImplementedError:
                 logger.info(
-                    "can't find generating function label %s."
-                    " The comb class is:\n%s",
+                    "can't find generating function for the rule %s -> %s. "
+                    "The rule was:\n%s",
                     self.get_label(rule.comb_class),
-                    rule.comb_class,
+                    tuple(self.get_label(child) for child in rule.children),
+                    rule,
                 )
                 x = var("x")
                 yield Eq(
                     self.get_function(rule.comb_class),
                     sympy.Function("NOTIMPLEMENTED")(x),
                 )
+
+    def get_initial_conditions(self, check: int = 6) -> List[Expr]:
+        """
+        Compute the initial conditions of the root class. It will use the
+        `count_objects_of_size` method if implemented, else resort to
+        the method on the `initial_conditions` method on `CombinatorialClass.
+        """
+        logger.info("Computing initial conditions")
+        try:
+            return [
+                sum(
+                    Number(self.count_objects_of_size(n=n, **parameters))
+                    * reduce(mul, [var(k) ** val for k, val in parameters.items()], 1)
+                    for parameters in self.root.possible_parameters(n)
+                )
+                for n in range(check + 1)
+            ]
+        except NotImplementedError as e:
+            logger.info(
+                "Reverting to generating objects from root for initial "
+                "conditions due to:\nNotImplementedError: %s",
+                e,
+            )
+        return self.root.initial_conditions(check)
 
     def get_genf(self, check: int = 6) -> Expr:
         """
@@ -243,21 +282,9 @@ class CombinatorialSpecification(
         """
         eqs = tuple(self.get_equations())
         root_func = self.get_function(self.root)
-        try:
-            logger.info("Computing initial conditions")
-            initial_conditions = [
-                self.count_objects_of_size(n=i) for i in range(check + 1)
-            ]
-        except NotImplementedError as e:
-            logger.info(
-                "Reverting to generating objects from root for initial "
-                "conditions due to:\nNotImplementedError: %s",
-                e,
-            )
-            initial_conditions = [
-                len(list(self.root.objects_of_size(i))) for i in range(check + 1)
-            ]
-        logger.info(maple_equations(root_func, initial_conditions, eqs,),)
+        logger.info("Computing initial conditions")
+        initial_conditions = self.get_initial_conditions(check)
+        logger.info(pretty_print_equations(root_func, initial_conditions, eqs))
         logger.info("Solving...")
         solutions = solve(
             eqs,
@@ -278,26 +305,24 @@ class CombinatorialSpecification(
                 return sympy.simplify(genf)
         raise IncorrectGeneratingFunctionError
 
-    def get_maple_equations(self, check: int = 6):
+    def get_maple_equations(self, check: int = 6) -> str:
+        """
+        Convert the systems of equations to version that can be copy pasted to maple.
+        """
         eqs = tuple(self.get_equations())
         root_func = self.get_function(self.root)
-        try:
-            logger.info("Computing initial conditions")
-            initial_conditions = [
-                self.count_objects_of_size(n=i) for i in range(check + 1)
-            ]
-        except NotImplementedError as e:
-            logger.info(
-                "Reverting to generating objects from root for initial "
-                "conditions due to:\nNotImplementedError: %s",
-                e,
-            )
-            initial_conditions = [
-                len(list(self.root.objects_of_size(i))) for i in range(check + 1)
-            ]
+        initial_conditions = self.get_initial_conditions(check)
         maple_eqs = maple_equations(root_func, initial_conditions, eqs)
-        logger.info(maple_eqs)
         return maple_eqs
+
+    def get_pretty_equations(self, check: int = 6) -> str:
+        """
+        Convert the systems of equations to a more readable format.
+        """
+        eqs = tuple(self.get_equations())
+        root_func = self.get_function(self.root)
+        initial_conditions = self.get_initial_conditions(check)
+        return pretty_print_equations(root_func, initial_conditions, eqs)
 
     def count_objects_of_size(self, n: int, **parameters) -> int:
         """
@@ -339,7 +364,11 @@ class CombinatorialSpecification(
         Raise an SanityCheckFailure error if it fails.
         """
         return all(
-            all(rule.sanity_check(n) for rule in self.rules_dict.values())
+            all(
+                rule.sanity_check(n, **parameters)
+                for rule in self.rules_dict.values()
+                for parameters in rule.comb_class.possible_parameters(n)
+            )
             for n in range(length + 1)
         )
 
@@ -429,7 +458,7 @@ class CombinatorialSpecification(
         }
 
     @classmethod
-    def from_dict(cls, d):
+    def from_dict(cls, d, expand_verified=False):
         """
         Return the specification with the dictionary outputter by the
         'to_jsonable' method
@@ -447,6 +476,7 @@ class CombinatorialSpecification(
                 [CombinatorialClass.from_dict(class_dict) for class_dict in eqv_path]
                 for eqv_path in d["eqv_paths"]
             ],
+            expand_verified=expand_verified,
         )